Android.mk: Remove unnecessary files for rpm build
Remove install.img and Androidx86-Installv28.5800.exe as it's not needed for rpm build. Build is still in progress so would be better if I can report whether the package succeeds to boot or not before merging.

Another suggestion is to prettify the lbuild success message when building with rpm so it looks similar to iso build. Also maybe follow iso output naming scheme for rpm package naming so people can expect addons and kernel bundled inside it